+++ src/main/java/appeng/items/storage/AbstractStorageCell.java
@@ -50,6 +50,7 @@ import net.minecraftforge.fml.relauncher.Side;
 import net.minecraftforge.fml.relauncher.SideOnly;
 import net.minecraftforge.items.IItemHandler;
 
+import java.util.Collections;
 import java.util.List;
 import java.util.Set;
 
@@ -119,7 +120,7 @@ public abstract class AbstractStorageCell extends AEBaseItem implements IBasicCe
 
     @Override
     public IItemHandler getConfigInventory(final ItemStack is) {
-        return new CellConfig(is);
+        return new CellConfig(is, Collections.singleton(this.getKeyType()));
     }
 
     @Override